Update mimemagic gem
The mimemagic project has run afoul of the GPL-2.0 license by including
GPL'ed bits in its codebase, but releasing under the MIT license. The
project's yanked all versions of the gem which contain the licensing
violation - < 0.3.5.0.

This updates to the re-licensed MIT version of the gem. This version
carries the caveat that host system must install the GPLed mime info
database.

References
----------

- https://github.com/rails/rails/issues/41757
- https://github.com/mimemagicrb/mimemagic/blob/master/README.md#dependencies
- https://github.com/rails/rails/issues/41750